Add files via upload
Part 1: Implement code to create and parse TFTP Packets (4 marks)
To start with, you are going to write code to create and parse TFTP Packets.  I have provided a skeleton implementation of a TftpPacket.java class file, with comments describing what you should do in each block of code.  Because this code is tricky, you need to write unit tests to make sure that you have got this part correct, before moving onto writing the code that sends and receives these packets.  I have provided a skeleton implementation of TftpPacketTest.java to help with this, with comments describing what you should do in each block of code.  You should implement ALL methods in both classes before moving onto the next part of the assignment.  When I wrote my implementation, I found bugs in my TftpPacket.java implementation with my unit tests, which I would have otherwise not easily found.  To compile and test:

javac -cp junit-platform-console-standalone-1.9.0.jar TftpPacketTest.java TftpPacket.java 
java -jar junit-platform-console-standalone-1.9.0.jar -cp . -c TftpPacketTest
You can download junit-platform-console-standalone-1.9.0.jar from here and place that jar file alongside your assignment java files.

Tips for TftpPacket.java:

I left createRRQ() implemented, in order for you to observe the general structure of these create methods.
For createACK() and createDATA(), the block number parameter is an integer, but you will be storing it into a byte array.  To do so, you'll have to cast the integer value stored in a variable named "block" to a byte value, as follows -- payload[1] = (byte)block; -- where foo is a byte array declared in the method.
I have used System.arraycopy to copy one byte array into another byte array, offset according to where I want the write to occur.  You could replace this method with a for loop, if you wanted.  You will be copying byte arrays in other parts of TftpPacket.java, so I left it in as an example.
For parse() I left parsing the RRQ intact, in order for you to observe the general structure you should use in the following if statements for other packet types.
When parsing DATA and ACK packets, you will need to extract the block number stored in each.  Java's bytes are signed values between -128 and +127, but we want to interpret each byte as an unsigned byte between 0 and 255.  To interpret data[1] as an unsigned byte, storing it in an integer, use something like block =  Byte.toUnsignedInt(data[1]);  The documentation for Byte.toUnsignedInt is here.
Valid block numbers are between 1 and 255.  If a server has more than 255 blocks to send, then the block number wraps to 1.  0 is not a valid block number.
Tips for TftpPacketTest.java
I left testGoodRRQ(), testGoodDATA(), testInvalidType(), and testEmptyBlock() intact, so that you could observe the general structure of these unit tests, and apply those concepts to other unit tests that you are to implement.

Part 2: Implement code in the client to send a RRQ, and code in the server to receive a RRQ (2 marks)
Next, you will start to implement TftpClient.java and TftpServer.java.  I have provided skeleton implementations of these class files, with comments describing what you should do in each block of code.  The server code will create a DatagramSocket to listen for RRQ packets to arrive.  When it receives a DatagramPacket, spawn a TftpServerWorker thread that (1) parses the DatagramPacket into a TftpPacket, (2) checks to see if the TftpPacket is valid -- not null, (3) is of RRQ type, and (4) obtains the filename.  The client code will create a DatagramSocket to send a RRQ packet, using the filename specified on the command line.

Print out the filename to a console in the server.  When this is working -- the TftpServer prints out the name of the file that you specified to TftpClient on the command line -- move onto the next part.

Part 3: Implement code in the server send DATA chunks (2 marks)
In the TftpServerWorker, create a new DatagramSocket that you will use to send packets to the client for this RRQ. You should expect an ACK packet for each DATA packet that you send; if you do not receive an ACK after one second, then retransmit the DATA packet.  If you try five times to send the DATA packet, and do not receive an ACK, then give up.  To set a timeout, use the setSoTimeout method on the DatagramSocket that you created.  This will cause the DatagramSocket's receive method to block for a defined length of time, throwing a SocketTimeoutException if the socket does not receive a packet in the defined time.

The server will try to open a FileInputStream using the filename it extracted from the RRQ.  If it cannot do so (because it caused a FileNotFoundException) then send an ERROR packet and return from the run method of the TftpServerWorker thread.  Otherwise, create a loop that reads 512 blocks of data from the file at a time.  For each block of data, wait for an ACK, as above.  Once you have got an ACK for the last packet, return from the TftpServerWorker thread.  In practice, you will not be able to finish this part of the assignment without simultaneously working on Part 4.

IMPORTANT: I left two break statements in the TftpServer.java file -- in the two infinite while loops, so that javac would compile the code.  Those two break statements, marked with /* XXX: delete the break, once you have implemented the loop */ should be deleted during the course of your implementation of this part.

Part 4: Implement code in the client to receive DATA packets (2 marks)
In the TftpClient, create a FileOutputStream that you will use to store the received file.  I suggest that you name the file after the file you requested; for example, if the request you made was for "cat.jpg", then name the output file "received-cat.jpg".  Set the soTimeout to 6 seconds, so that the client can exit if the server fails to deliver a DATA packet after 6 seconds of trying.

Inside an infinite loop, read DatagramPackets from the socket you sent the RRQ on.  If you cannot parse the packet (TftpPacket::parse returns null) then use the "continue" keyword to loop again without executing the remainder of the code in the loop. If you receive an ERROR packet, then abort after printing the error message.  If the packet is not a DATA packet, then use the continue keyword to loop again.  Check the block number in the DATA packet.

If the block number is exactly what you expect, then write the block to disk, send an ack, and increment the block number (using the TftpPacket.nextBlock utility function to handle wrap-around from 255 back to 1).  If the block was less than 512 bytes, then exit after you've sent the ACK!

If the block number in the DATA packet is one less than what you expected, then send an ack without writing the block to disk.  In this case, you've received a duplicate of the last DATA packet, perhaps because the ACK you sent was lost in the network.